+++ src/main/java/com/excilys/formation/cdb/paginator/core/LimitValue.java
@@ -1,17 +1,17 @@
 package com.excilys.formation.cdb.paginator.core;
 
 public enum LimitValue {
-    TEN(10),
-    TWENTY(20),
-    ONE_HUNDRED(100);
+    TEN(10l),
+    TWENTY(20l),
+    ONE_HUNDRED(100l);
 
-    private final int value;
+    private final Long value;
 
-    LimitValue(final int newValue) {
+    LimitValue(final Long newValue) {
         value = newValue;
     }
 
-    public int getValue() {
+    public Long getValue() {
         return value;
     }
 }
